Move from `htm/babel` to `babel-plugin-htm`
