+++ .gitignore
@@ -1,6 +1,5 @@
-/node_modules
-/package-lock.json
+node_modules
+package-lock.json
 /preact
-/babel
-/dist
-/yarn.lock
+dist
+yarn.lock
